+++ lib/ex_bashkit/native.ex
@@ -33,4 +33,6 @@ defmodule ExBashkit.Native do
   # Sessions (persistent, stateful interpreters held as a resource).
   def session_new(_env, _cwd, _username, _hostname), do: :erlang.nif_error(:nif_not_loaded)
   def session_exec(_session, _script), do: :erlang.nif_error(:nif_not_loaded)
+  def session_read_file(_session, _path), do: :erlang.nif_error(:nif_not_loaded)
+  def session_write_file(_session, _path, _content), do: :erlang.nif_error(:nif_not_loaded)
 end